Restructure packaging for openbanking
COAND-497
