+++ dotpay/src/main/java/com/adyen/checkout/dotpay/DotpayComponent.kt
@@ -14,7 +14,7 @@ import com.adyen.checkout.components.base.ComponentEventHandler
 import com.adyen.checkout.components.model.payments.request.DotpayPaymentMethod
 import com.adyen.checkout.components.util.PaymentMethodTypes
 import com.adyen.checkout.dotpay.DotpayComponent.Companion.PROVIDER
-import com.adyen.checkout.dotpay.interal.provider.DotpayComponentProvider
+import com.adyen.checkout.dotpay.internal.provider.DotpayComponentProvider
 import com.adyen.checkout.issuerlist.IssuerListComponent
 import com.adyen.checkout.issuerlist.internal.ui.IssuerListDelegate
