+++ openbanking/src/main/java/com/adyen/checkout/openbanking/OpenBankingComponent.kt
@@ -16,6 +16,7 @@ import com.adyen.checkout.components.util.PaymentMethodTypes
 import com.adyen.checkout.issuerlist.IssuerListComponent
 import com.adyen.checkout.issuerlist.internal.ui.IssuerListDelegate
 import com.adyen.checkout.openbanking.OpenBankingComponent.Companion.PROVIDER
+import com.adyen.checkout.openbanking.internal.provider.OpenBankingComponentProvider
 
 /**
  * Component should not be instantiated directly. Instead use the [PROVIDER] object.
